+++ CookieData/Model/GameAccountModel.cs
@@ -5,7 +5,7 @@ namespace CookieData.Model
     public class GameAccountModel
     {
         public int Id { get; set; }
-        public long Networth { get; set; }
+        public double Networth { get; set; }
         public long Cookies { get; set; }
         public long Clicks { get; set; }
         public ClickUpgradeModel ClickUpgrade { get; set; }
